5755 - refactor i18n initialisation
