+++ src/server/controller/user/getManyExport.ts
@@ -7,7 +7,7 @@ import { Objects } from 'utils/objects'
 
 import { UserRepository, UsersGetManyProps } from 'server/db/repository/public/user'
 import { UserQueryParams } from 'server/db/repository/public/user/UserQueryParams'
-import { I18nUtils } from 'server/utils'
+import { I18n } from 'server/utils'
 
 type Props = UsersGetManyProps & {
   lang: Lang
@@ -71,7 +71,7 @@ export const getManyExport = async (props: Props): Promise<Returned> => {
   const { lang } = props
   const { query, queryParams } = UserRepository.buildGetManyExportQuery(props)
 
-  const i18n = await I18nUtils.get({ lang })
+  const i18n = await I18n.get({ lang })
 
   const rowTransformer = (rawRow: RawExportRow): Record<string, string> => {
     const { role: rawRole, ...userFields } = rawRow
