+++ src/server/service/pdfReport/generate.ts
@@ -8,7 +8,7 @@ import { CycleName } from 'meta/assessment/cycle'
 import { Lang } from 'meta/lang'
 import { Promises } from 'utils/promises'
 
-import { I18nUtils, ProcessEnv } from 'server/utils'
+import { I18n, ProcessEnv } from 'server/utils'
 import { Logger } from 'server/utils/logger'
 
 type Props = {
@@ -52,7 +52,7 @@ const browserOptions: LaunchOptions = debug ? { headless: false, defaultViewport
 
 export const generate = async (props: Props): ReturnType<Page['pdf']> => {
   const { appUri, assessmentName, cookies, countryIso, cycleName, lang, onlyTables } = { ...defaultProps, ...props }
-  const { t } = await I18nUtils.get({ lang })
+  const { t } = await I18n.get({ lang })
   const browser = await puppeteer.launch(browserOptions)
   const page = await browser.newPage()
