+++ src/server/utils/i18n.ts
@@ -1,5 +1,5 @@
-import { createI18nPromise } from 'i18n/i18nFactory'
-import { TFunction } from 'i18next'
+import { createParams } from 'i18n/i18nConfig'
+import i18next, { i18n as i18nType } from 'i18next'
 
 import { Lang } from 'meta/lang'
 import { User } from 'meta/user/user'
@@ -9,7 +9,10 @@ type Props = {
   lang?: Lang
 }
 
-const get = ({ lang, user }: Props = {}): Promise<{ language: Lang; t: TFunction }> =>
-  createI18nPromise(user?.props.lang ?? lang ?? Lang.en)
+const get = async ({ lang, user }: Props = {}): Promise<i18nType> => {
+  const instance = i18next.createInstance(createParams(user?.props.lang ?? lang ?? Lang.en))
+  await instance.init()
+  return instance
+}
 
-export const I18nUtils = { get }
+export const I18n = { get }
