+++ src/server/utils/index.ts
@@ -1,3 +1,3 @@
-export { I18nUtils } from './i18n'
+export { I18n } from './i18n'
 export { ProcessEnv } from './processEnv'
 export { Requests } from './requests'
